s23 tests and code formatting

+++ cmd/bitcoin2utxoset/bitcoin/headers.go
@@ -100,7 +100,7 @@ func (in *IndexDB) WriteHeadersToFile(outputDir string, heightHint int) (*utxope
 	bufferedWriter := bufio.NewWriter(io.MultiWriter(file, hasher))
 	defer bufferedWriter.Flush()
 
-	header, err := utxopersister.BuildHeaderBytes("U-H-1.0", bestBlock.Hash, uint32(bestBlock.Height), bestBlock.BlockHeader.HashPrevBlock)
+	header, err := utxopersister.BuildHeaderBytes("U-H-1.0", bestBlock.Hash, bestBlock.Height, bestBlock.BlockHeader.HashPrevBlock)
 	if err != nil {
 		return nil, errors.NewProcessingError("Couldn't build UTXO set header", err)
 	}
